Se corrigen errores del merge
