+++ public/index.html
@@ -13,7 +13,7 @@
     <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
 
-    <title>UNAHUR-Covid</title>
+    <title>UNaHur-Covid</title>
 
     <link
       rel="stylesheet"
