+++ src/App.js
@@ -1,12 +1,11 @@
 import { Box, Container } from '@material-ui/core';
 import { Route, BrowserRouter as Router, Switch } from 'react-router-dom';
 
-import AltaActividad from './components/actividades/AltaActividad';
-import Header from './components/Header';
-import Home from './components/Home';
+import AltaModificacionActividad from './components/actividades/AltaModificacionActividad';
+import AltaModificacionEspacio from './components/espacios/AltaModificacionEspacio';
+import Header from './components/ui/Header';
 import ListadoActividades from './components/actividades/ListadoActividades';
-import PantallaEspacios from './components/PantallaEspacios';
-import Espacio from './components/Espacio';
+import ListadoEspacios from './components/espacios/ListadoEspacios';
 
 export default function App() {
   return (
@@ -15,26 +14,26 @@ export default function App() {
       <Box my={4}>
         <Router>
           <Switch>
+            <Route exact path="/">
+              <ListadoEspacios />
+            </Route>
+            <Route exact path="/espacios">
+              <ListadoEspacios />
+            </Route>
             <Route path="/espacios/nuevo">
-              <Espacio titulo={'Carga de espacios'} />
+              <AltaModificacionEspacio titulo={'Carga de espacios'} />
             </Route>
             <Route path="/espacios/:id">
-              <Espacio titulo={'Modificar un espacio'} />
-            </Route>
-            <Route exact path="/">
-              <Home />
-            </Route>
-            <Route path="/espacios">
-              <PantallaEspacios />
+              <AltaModificacionEspacio titulo={'Modificar un espacio'} />
             </Route>
             <Route exact path="/actividades">
               <ListadoActividades />
             </Route>
             <Route path="/actividades/nueva">
-              <AltaActividad titulo={'Carga de actividades'} />
+              <AltaModificacionActividad titulo={'Carga de actividades'} />
             </Route>
             <Route path="/actividades/:id">
-              <AltaActividad titulo={'Editar actividad'} />
+              <AltaModificacionActividad titulo={'Editar actividad'} />
             </Route>
           </Switch>
         </Router>
