--- src/components/Home.js
@@ -1,11 +1,0 @@
-import { Box, Typography } from '@material-ui/core';
-
-export default function Home() {
-  return (
-    <Box mt={8}>
-      <Typography variant="h4" color="primary">
-        Inicio
-      </Typography>
-    </Box>
-  );
-}
